Editing

+++ Accounts/DataLayer/AdoNet/ProductDao.vb
@@ -64,7 +64,7 @@ Namespace DataLayer.AdoNet
             New Product() With {
             .Active = Extensions.AsBool(reader("Active")),
             .BarCode = Extensions.AsString(reader("Barcode")),
-            .BaseUnitIdNo = Extensions.AsInt(Of Int16)(reader("BaseUnitIdNo")),
+            .BaseUnitIdNo = Extensions.AsInt(Of Int32)(reader("BaseUnitIdNo")),
             .CategoryIdNo = Extensions.AsInt(Of Int16)(reader("CategoryIdNo")),
             .GTIN = Extensions.AsString(reader("GTIN")),
             .IdNo = Extensions.AsId(Of Int32)(reader("IdNo")),
@@ -144,7 +144,7 @@ Namespace DataLayer.AdoNet
             .BaseQty = Extensions.AsInt(Of Int16)(reader("BaseQty")),
             .IdNo = Extensions.AsId(Of Int32)(reader("IdNo")),
             .Multiplier = Extensions.AsInt(Of Int16)(reader("Multiplier")),
-            .UnitIdNo = Extensions.AsInt(Of Int16)(reader("UnitIdNo"))
+            .UnitIdNo = Extensions.AsInt(Of Int32)(reader("UnitIdNo"))
            }
 
     End Class
