+++ Accounts/PresentationLayer/Presenters/ProductPresenter.vb
@@ -33,7 +33,7 @@ Namespace PresentationLayer.Presenters
             CreateDataTable(DtProductUnitInsertTable, {{"BaseQty", GetType(Int16)},
                                  {"Multiplier", GetType(Int16)},
                                  {"ProductIdNo", GetType(Int32)},
-                                 {"UnitIdNo", GetType(Int16)}
+                                 {"UnitIdNo", GetType(Int32)}
                                  })
 
 
@@ -41,7 +41,7 @@ Namespace PresentationLayer.Presenters
                                              {"IdNo", GetType(Int32)},
                                              {"Multiplier", GetType(Int16)},
                                              {"ProductIdNo", GetType(Int32)},
-                                             {"UnitIdNo", GetType(Int16)}
+                                             {"UnitIdNo", GetType(Int32)}
                                             })
 
         End Sub
