+++ Accounts/PresentationLayer/Presenters/PurchasePresenter.vb
@@ -29,7 +29,7 @@ Namespace PresentationLayer.Presenters
             DtInsertTable.Columns.Add("PurchaseIdNo", GetType(Int32))
             DtInsertTable.Columns.Add("Quantity", GetType(Int16))
             DtInsertTable.Columns.Add("Sequence", GetType(Int16))
-            DtInsertTable.Columns.Add("UnitIdNo", GetType(Int16))
+            DtInsertTable.Columns.Add("UnitIdNo", GetType(Int32))
             DtInsertTable.Columns.Add("VatAmount", GetType(Decimal))
             DtInsertTable.Columns.Add("VatPercent", GetType(Decimal))
 
@@ -41,7 +41,7 @@ Namespace PresentationLayer.Presenters
             DtUpdateTable.Columns.Add("PurchaseIdNo", GetType(Int32))
             DtUpdateTable.Columns.Add("Quantity", GetType(Int16))
             DtUpdateTable.Columns.Add("Sequence", GetType(Int16))
-            DtUpdateTable.Columns.Add("UnitIdNo", GetType(Int16))
+            DtUpdateTable.Columns.Add("UnitIdNo", GetType(Int32))
             DtUpdateTable.Columns.Add("VatAmount", GetType(Decimal))
             DtUpdateTable.Columns.Add("VatPercent", GetType(Decimal))
